requirements.txt
absl-py==0.6.1
astor==0.7.1
breadability==0.1.20
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
dill==0.2.8.2
docopt==0.6.2
future==0.16.0
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
idna==2.7
joblib==0.13.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
llvmlite==0.24.0
lxml==4.2.5
Mako==1.0.7
networkx==2.2
nltk==3.3
numba==0.39.0
numpy==1.15.2
pandas==0.23.4
protobuf==3.6.1
pyrouge==0.1.3
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rouge==0.3.1
scikit-learn==0.21.2
scipy==1.1.0
six==1.11.0
sumy==0.7.0
tensorboard==1.12.1
termcolor==1.1.0
torch==0.4.1
torchtext==0.3.1
tqdm==4.26.0
urllib3==1.23
Werkzeug==0.14.1
